separate internal and integration test
